% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brendaDb.R
\docType{data}
\name{acronyms}
\alias{acronyms}
\title{Information fields and their corresponding acronyms.}
\format{A data.frame with 40 rows and 2 columns.
\itemize{
\item field: field names in the BRENDA text file (all uppercase letters separated
by underscores). This could be used in the \code{fields} argument of the
\code{QueryBrenda()} function.
\item acronym: acronyms used by the BRENDA file to indicate the fields.
}}
\usage{
data(acronyms)
}
\description{
The RData file is generated by reading the entire BRENDA text file with
\code{ReadBrenda()}, and getting unique values in the \code{field} column. The acronyms
are extracted by the regex \verb{^[A-Z_05]+}.
}
\keyword{internal}
