% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocyc.R
\name{BiocycPathwayEnzymes}
\alias{BiocycPathwayEnzymes}
\title{Get all EC numbers involved in a BioCyc pathway.}
\usage{
BiocycPathwayEnzymes(org.id = "HUMAN", pathway)
}
\arguments{
\item{org.id}{The identifier for the organism database in BioCyc, e.g. ECOLI,
HUMAN, META, AFER243159}

\item{pathway}{A case-sensitive pathway object identifier,
e.g. PWY66-400, LYSINE-DEG1-PWY.}
}
\value{
If the pathway is found, returns a tibble with columns \code{Reaction} and
\code{EC}, where \code{Reaction} is the reaction IDs found in the pathway. Returns NULL
if the pathway ID is not found.
}
\description{
Get all EC numbers involved in a BioCyc pathway.
}
\examples{
BiocycPathwayEnzymes("HUMAN", "PWY66-400")

}
