% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocyc.R
\name{BiocycPathwayGenes}
\alias{BiocycPathwayGenes}
\title{Get all genes involved in a BioCyc pathway.}
\usage{
BiocycPathwayGenes(org.id = "HUMAN", pathway)
}
\arguments{
\item{org.id}{The identifier for the organism database in BioCyc, e.g. ECOLI,
HUMAN, META, AFER243159}

\item{pathway}{A case-sensitive pathway object identifier,
e.g. PWY66-400, LYSINE-DEG1-PWY.}
}
\value{
If the pathway is found, returns a tibble with columns \code{BiocycGene},
\code{BiocycProtein}, \code{Symbol} and \code{Ensembl}, where \code{BiocycGene} and
\code{BiocycProtein} are the gene and protein IDs in the BioCyc database,
respectively. Returns NULL if the pathway ID is not found.
}
\description{
Get all genes involved in a BioCyc pathway.
}
\examples{
BiocycPathwayGenes("HUMAN", "PWY66-400")
BiocycPathwayGenes("HUMAN", "TRYPTOPHAN-DEGRADATION-1")

}
