% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ConfigBPCores}
\alias{ConfigBPCores}
\title{Configure the number of cores used by BiocParallel.}
\usage{
ConfigBPCores(n.core = 0)
}
\arguments{
\item{n.core}{Integer specifying the number of cores to use. Default is 0,
which would result in using all available cores.}
}
\value{
The back-end of type \code{bpparamClass}.
}
\description{
Configure the number of cores used by BiocParallel.
}
\examples{
brendaDb:::ConfigBPCores(2)

}
\keyword{internal}
