% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{DownloadBrenda}
\alias{DownloadBrenda}
\title{Download and unzip the BRENDA text file.}
\usage{
DownloadBrenda(force.download = FALSE)
}
\arguments{
\item{force.download}{Boolean value. If TRUE, ignore the cache and force
re-download of the BRENDA text file. Default is FALSE.}
}
\value{
A string of the path to the downloaded BRENDA text file.
}
\description{
By default, the function downloads a zipped BRENDA text file to
a local cache directory, and extracts a \code{brenda_download.txt} file.
}
\examples{
\dontrun{DownloadBrenda()}

}
