% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brendaEntry.R
\name{InitBrendaEntry}
\alias{InitBrendaEntry}
\alias{is.brenda.entry}
\title{Create a \code{brendaEntry} object.}
\usage{
InitBrendaEntry(EC, protein = NA, systematic.name = NA,
  recommended.name = NA, synonyms = NA, reaction = NA,
  reaction.type = NA, substrate.product = NA,
  natural.substrate.product = NA, cofactor = NA, metals.ions = NA,
  inhibitors = NA, activating.compound = NA, km.value = NA,
  turnover.number = NA, ki.value = NA, pi.value = NA,
  ph.optimum = NA, ph.range = NA, temperature.optimum = NA,
  temperature.range = NA, specific.activity = NA, ic50 = NA,
  source.tissue = NA, localization = NA, general.stability = NA,
  storage.stability = NA, ph.stability = NA,
  organic.solvent.stability = NA, oxidation.stability = NA,
  temperature.stability = NA, purification = NA, cloned = NA,
  engineering = NA, renatured = NA, application = NA,
  molecular.weight = NA, subunits = NA,
  posttranslational.modification = NA, cystallization = NA,
  bibliography = NA)

is.brenda.entry(x)
}
\arguments{
\item{EC}{A string indicating EC number of the enzyme.}

\item{x}{Any object.}
}
\description{
The list should contain 6 sublists: \code{nomenclature},
\code{interactions}, \code{parameters}, \code{molecular}, \code{stability}, and \code{bibliography}.
All sublists should be empty apart from the \code{EC} term in \code{nomenclature}.
}
\examples{
brendaDb:::InitBrendaEntry("1.1.1.100")
}
