% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{long_to_wide}
\alias{long_to_wide}
\title{Convert vector of lines to matrix.}
\usage{
long_to_wide(lines)
}
\arguments{
\item{fin}{The output vector<string> from \code{read_brenda_file}.}
}
\value{
A vector<vector<string>> containing information about the EC entries.
}
\description{
For each EC entry, split the annotations into three columns:
\itemize{
\item ID: EC number, e.g. 1.1.1.1
\item field: the content of the information, e.g. protein, localization
\item description: everything else
}
}
