% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-bibliography.R
\name{ParseApplication}
\alias{ParseApplication}
\title{Parse a "APPLICATION" entry.}
\usage{
ParseApplication(description)
}
\arguments{
\item{description}{The description string in a "APPLICATION" entry.}
}
\value{
A \code{data.table} with columns: proteinID, description, fieldInfo,
commentary and refID. The description column is the application, and the
fieldInfo column should be all NAs.
}
\description{
Expand the string into a \code{data.table}.
}
\examples{
x <- paste0(
"AP\\t#13# biotechnology (#13# possible usage in bioindustrial\\n\\t",
"processes and as biosensor <126>) <126>\\n",
"AP\\t#6,10,13,15,28,29,34,36,39,41,43,88,92,118,123,139,140,143,144,\\n\\t",
"151,152# synthesis (#43# enzyme can be used in preparative... <1>)",
"<114,131,132,133,134,137,169,179,187,219,232,234,240,242,245,246,248\\n\\t",
"250,251,253,254,261,262,263,264,266,269,272"
)
brendaDb:::ParseApplication(x)
}
