% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-helper.R
\name{ParseCommentary}
\alias{ParseCommentary}
\title{Parse commentaries inside the description.}
\usage{
ParseCommentary(description)
}
\arguments{
\item{description}{A BRENDA description entry with commentaries.}
}
\value{
A list of commentaries.
}
\description{
Parse the commentaries inside parentheses in the given string.
}
\examples{
brendaDb:::ParseCommentary("Cavia porcellus   (#1# SULT1A2 <1,2,6,7>)")

}
