% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-helper.R
\name{ParseGeneric}
\alias{ParseGeneric}
\title{Generic parser for a list of descriptions.}
\usage{
ParseGeneric(des.list)
}
\arguments{
\item{des.list}{A list of descriptions separated by \code{SeparateSubentries}.}
}
\value{
A \code{data.table} with columns: proteinID, description, fieldInfo,
commentary, and refID
}
\description{
Descriptions are generally structured as the following:
\itemize{
\item Protein information is included in '#'...#',
\item Literature citations are in '<...>',
\item Commentaries in '(...)', and
\item field-special information in '{...}'.
}
}
