% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-helper.R
\name{ParseNoDescription}
\alias{ParseNoDescription}
\title{Generic parser for a description string without extracted values.}
\usage{
ParseNoDescription(description, acronym)
}
\arguments{
\item{description}{A description string from one of the entries.}

\item{acronym}{The acronym of the field. Can be found with \code{\link[=ShowFields]{ShowFields()}}.}
}
\value{
A \code{tibble} with columns: proteinID, description and refID.
}
\description{
This parser works for fields \code{storage.stability},
\code{general.stability}, \code{oxidation.stability}, \code{cloned}, \code{purification},
\code{crystallization} and \code{renatured}.

These fields in BRENDA don't have extracted values - the commentary itself is
the extracted value.
}
