% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-parameters.R
\name{ParsePhOptimum}
\alias{ParsePhOptimum}
\title{Parse a "PH_OPTIMUM" entry.}
\usage{
ParsePhOptimum(description)
}
\arguments{
\item{description}{The description string in a "PH_OPTIMUM" entry.}
}
\value{
A \code{data.table} with columns: proteinID, description, fieldInfo,
commentary and refID. The description column is the extracted pH optimum
value/range, and the fieldInfo column should all be NAs.
}
\description{
Expand the string into a \code{data.table}.
}
\examples{
x <- paste0(
"PHO\\t#10# 8.3 (#10# alcohol dehydrogenase IV <87>) <87>\\n",
"PHO\\t#9,68,78,132# 11 (#78# oxidation of ethanol, pyrazole-insensitive\\n\\t",
"enzyme <24>; #68# ethanol oxidation, enzyme form ADH-2 and ADH-3 <60>;\\n\\t",
"#9# oxidation of octanol <49>; #132# optimally active with ethanol and\\n\\t",
"1-propanol at pH 11.0 with 3 M KCl <237>) <24,49,60,237>\\n")
brendaDb:::ParsePhOptimum(x)
}
