% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-parameters.R
\name{ParsePhRange}
\alias{ParsePhRange}
\title{Parse a "PH_RANGE" entry.}
\usage{
ParsePhRange(description)
}
\arguments{
\item{description}{The description string in a "PH_RANGE" entry.}
}
\value{
A \code{data.table} with four columns:
id, pH_range, commentary and references.
}
\description{
Expand the string into a \code{data.table}.
}
