% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-reaction.R
\name{ParseReaction}
\alias{ParseReaction}
\title{Parser for the description strings of certain reaction-related fields.}
\usage{
ParseReaction(description, acronym)
}
\arguments{
\item{description}{A description string from one of the entries.}

\item{acronym}{The acronym of the field. Can be found with \code{\link[=ShowFields]{ShowFields()}}.}
}
\value{
A \code{tibble} with columns: proteinID, substrate, product,
commentarySubstrate, commentaryProduct, fieldInfo and refID
}
\description{
This parser works for the fields \code{substrate.product} and
\code{natural.substrate.product}.
}
\details{
The reversibility of the reactions are wrapped in {}; substrates
and products are separated by "="; commentaries on substrates are wrapped in
"()", and commentaries on products are wrapped in "||".
}
\keyword{internal}
