% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-nomenclature.R
\name{ParseRecommendedName}
\alias{ParseRecommendedName}
\title{Parse a "RECOMMENDED_NAME" entry into a string.}
\usage{
ParseRecommendedName(description)
}
\arguments{
\item{description}{The description string in a "RECOMMENDED_NAME" entry.}
}
\value{
A string to fill into the \code{recommended.name} field in
\code{brenda.nomenclature}.
}
\description{
Remove useless characters in the description.
}
\examples{
x <- "RN\\tD-arabinose 1-dehydrogenase (NAD+)"
brendaDb:::ParseRecommendedName(x)
}
