% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-nomenclature.R
\name{ParseSystematicName}
\alias{ParseSystematicName}
\title{Parse a "SYSTEMATIC_NAME" entry into a string.}
\usage{
ParseSystematicName(description)
}
\arguments{
\item{description}{The description string in a "SYSTEMATIC_NAME" entry.}
}
\value{
A string to fill into the \code{systematic.name} field in
\code{brenda.nomenclature}.
}
\description{
Remove useless characters in the description.
}
\examples{
x <- "SN\talcohol:NAD+ oxidoreductase"
brendaDb:::ParseSystematicName(x)
}
\keyword{internal}
