% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{PrettyPrintBrendaEntry}
\alias{PrettyPrintBrendaEntry}
\title{Print a brenda.entry in a tree view.}
\usage{
PrettyPrintBrendaEntry(x, index, tail.idx, depth)
}
\arguments{
\item{x}{A \code{brenda.entry} object.}

\item{index}{A string of the name of the sublist.}

\item{tail.idx}{A string showing the last element in the entry. This is for
printing a different character in the tree.}

\item{depth}{Int, showing the depth of the element.}
}
\description{
Print a brenda.entry in a tree view.
}
