% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{print.brenda.entries}
\alias{print.brenda.entries}
\title{Show the number of regular and transferred/deleted brenda.entry objects
in the brenda.entries list.}
\usage{
\method{print}{brenda.entries}(x, ..., verbose = F)
}
\arguments{
\item{x}{A brenda.entries list returned by \code{\link[=QueryBrenda]{QueryBrenda()}}.}

\item{...}{Other arguments passed to the generic function.}

\item{verbose}{Boolean; if TRUE, print tree views of each brenda.query object.}
}
\value{
Nothing; print summary information to the terminal.
}
\description{
Show the number of regular and transferred/deleted brenda.entry objects
in the brenda.entries list.
}
