% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{print.brenda.entry}
\alias{print.brenda.entry}
\title{Show the non-empty fields in the query result.}
\usage{
\method{print}{brenda.entry}(x, full.output = FALSE, ...)
}
\arguments{
\item{x}{A brenda.entry object (elements in the list returned by the function
\code{\link[=QueryBrenda]{QueryBrenda()}}).}

\item{full.output}{A boolean default to FALSE. If TRUE, include all entries
even if they are empty (NA or 0 rows).}

\item{...}{Other arguments passed to the generic function.}
}
\value{
Nothing; print object information to the terminal.
}
\description{
For details, see \code{\link[=PrettyPrintBrendaEntry]{PrettyPrintBrendaEntry()}}.
}
