% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{PrintTreeHelper}
\alias{PrintTreeHelper}
\title{Helper function for printing the tree structure with correct whitespace.}
\usage{
PrintTreeHelper(index, tail.idx, depth)
}
\arguments{
\item{index}{A string of the name of the sublist.}

\item{tail.idx}{A string showing the last element in the entry. This is for
printing a different character in the tree.}

\item{depth}{Int, showing the depth of the element.}
}
\value{
Nothing; prints tree structure to the terminal.
}
\description{
Helper function for printing the tree structure with correct whitespace.
}
