% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_brenda}
\alias{read_brenda}
\title{Read BRENDA text file into matrix of strings.}
\usage{
read_brenda(filepath)
}
\arguments{
\item{filepath}{A string indicating the path to the text file.}
}
\value{
A vector<vector<string>> containing information about the EC entries. In R
this is a list of lists.
}
\description{
For each EC entry, split the annotations into three columns:
\itemize{
\item ID: EC number, e.g. 1.1.1.1
\item field: the content of the information, e.g. protein, localization
\item description: everything else
}
}
