% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadBrendaFile}
\alias{ReadBrendaFile}
\title{Read raw BRENDA text file.}
\usage{
ReadBrendaFile(filepath)
}
\arguments{
\item{filepath}{A string indicating the path to the text file.}
}
\value{
A vector\if{html}{\out{<string>}} with each element being a line in the file.
}
\description{
Read file into buffer, and load all non-empty lines. Comment lines
(starting with *) are skipped. The text file should be downloaded from
https://www.brenda-enzymes.org/download_brenda_without_registration.php
}
