% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{SelectOrganism}
\alias{SelectOrganism}
\title{Select a subset of organisms in the brenda.query object.}
\usage{
SelectOrganism(query, org.id)
}
\arguments{
\item{query}{A brenda.entry object from \code{\link[=QueryBrendaBase]{QueryBrendaBase()}}.}

\item{org.id}{A named character vector with values as \code{proteinID}s.}
}
\value{
A subset of the given brenda.query object with the given organisms
selected.
}
\description{
Select a subset of organisms in the brenda.query object.
}
\seealso{
\code{\link[=QueryBrendaBase]{QueryBrendaBase()}}
}
