% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-helper.R
\name{SeparateSubentries}
\alias{SeparateSubentries}
\title{Preprocessing of entry description.}
\usage{
SeparateSubentries(description, acronym = NA)
}
\arguments{
\item{description}{The \code{description} string of the field.}

\item{acronym}{The acronym of the field. Can be found with \code{\link[=ShowFields]{ShowFields()}}.}
}
\value{
A list of strings with each subentry as an element.
}
\description{
Separate subentries and strip unnecessary whitespace.
}
\examples{
x <- "SN\\talcohol:NAD+ oxidoreductase"
brendaDb:::SeparateSubentries(x, "SN")
}
\keyword{internal}
