% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{show.fields}
\alias{show.fields}
\title{Show all unique BRENDA fields and their corresponding acronyms.}
\usage{
show.fields(df)
}
\arguments{
\item{df}{A data.frame with columns "field" and "description"}
}
\value{
A data.frame with columns "field" and "acronym".
}
\description{
Show all unique BRENDA fields and their corresponding acronyms.
}
