% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{ShowFields}
\alias{ShowFields}
\title{Show all unique BRENDA fields and their corresponding acronyms.}
\usage{
ShowFields(df)
}
\arguments{
\item{df}{A data.frame with columns "field" and "description"}
}
\value{
A data.frame with columns "field" and "acronym".
}
\description{
Show all unique BRENDA fields and their corresponding acronyms.
}
\examples{
df <- ReadBrenda(system.file("extdata", "brenda_download_test.txt",
                          package = "brendaDb"))
ShowFields(df)

}
