% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{summary.brenda.entries}
\alias{summary.brenda.entries}
\title{Show the number of regular and transferred/deleted brenda.entry objects
in the brenda.entries list.}
\usage{
\method{summary}{brenda.entries}(x, verbose = F)
}
\arguments{
\item{x}{A brenda.entries list returned by \code{\link[=QueryBrenda]{QueryBrenda()}}.}
}
\description{
Show the number of regular and transferred/deleted brenda.entry objects
in the brenda.entries list.
}
