% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{summary.brenda.entry}
\alias{summary.brenda.entry}
\title{Show the non-empty fields in the query result.}
\usage{
\method{summary}{brenda.entry}(x)
}
\arguments{
\item{x}{A brenda.entry object (elements in the list returned by \code{\link[=QueryBrenda]{QueryBrenda()}}).}
}
\description{
For details, see \code{\link[=PrettyPrintBrendaEntry]{PrettyPrintBrendaEntry()}}.
}
