% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{summary}
\alias{summary}
\title{Generic function for showing summaries of brenda.entry objects.}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{A \code{brenda.entry} or \code{brenda.entries} object.}

\item{...}{Other arguments passed to the generic function.}
}
\description{
Generic function for showing summaries of brenda.entry objects.
}
\seealso{
\code{\link[=summary.brenda.entries]{summary.brenda.entries()}} \code{\link[=summary.brenda.entry]{summary.brenda.entry()}}
}
