% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brendaDb.R
\docType{data}
\name{acronyms}
\alias{acronyms}
\title{Information fields and their corresponding acronyms}
\format{A data.frame with 39 rows and 2 variables}
\usage{
data(acronyms)
}
\description{
Information fields and their corresponding acronyms
}
\keyword{internal}
