% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocyc.R
\name{BiocycPathwayEnzymes}
\alias{BiocycPathwayEnzymes}
\title{Get all EC numbers involved in a BioCyc pathway.}
\usage{
BiocycPathwayEnzymes(org.id = "HUMAN", pathway, sleep = 0)
}
\arguments{
\item{org.id}{The identifier for the organism database in BioCyc, e.g. ECOLI,
HUMAN, META, AFER243159}

\item{pathway}{A case-sensitive pathway object identifier,
e.g. PWY66-400, LYSINE-DEG1-PWY.}

\item{sleep}{Optional, default is 0. If set to a non-zero value, pause for
the specified seconds after retrieving each reaction. Increase this if the
default fails (probably limited on the BioCyc API side).}
}
\value{
If the pathway is found, returns a tibble with columns
- \code{RxnID}: reaction IDs found in the pathway.
- \code{EC}: EC number of the enzyme catalyzing the reaction.
- \code{ReactionDirection}: direction of the reaction.
- \code{LHS}: left-hand-side of the reaction.
- \code{RHS}: right-hand-side of the reaction.
Different compounds in the reactions are separated by " + " in columns \code{LHS}
and \code{RHS}. The function returns NULL if the pathway ID is not found.
}
\description{
Get all EC numbers involved in a BioCyc pathway.
}
\examples{
BiocycPathwayEnzymes("HUMAN", "PWY66-400")

}
