% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brendaDb.R
\docType{package}
\name{brendaDb-package}
\alias{brendaDb}
\alias{brendaDb-package}
\title{brendaDb: the BRENDA enzyme database}
\description{
The main aims of brendaDb include:
\itemize{
\item Make it easier to import and analyze enzyme information in R
\item Generate enzyme activity profiles
}
}
