% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brendaEntry.R
\name{InitBrendaEntry}
\alias{InitBrendaEntry}
\alias{is.brenda.entry}
\title{Create a \code{brendaEntry} object.}
\usage{
InitBrendaEntry(EC)

is.brenda.entry(x)
}
\arguments{
\item{EC}{A string indicating EC number of the enzyme.}

\item{x}{Any object.}
}
\description{
The list should contain 6 sublists: \code{nomenclature},
\code{interactions}, \code{parameters}, \code{molecular}, \code{stability}, and \code{bibliography}.
All sublists should be empty apart from the \code{EC} term in \code{nomenclature}.
}
\examples{
brendaDb:::InitBrendaEntry("1.1.1.100")
}
