% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-parameters.R
\name{ParsePhRange}
\alias{ParsePhRange}
\title{Parse a "PH_RANGE" entry.}
\usage{
ParsePhRange(description)
}
\arguments{
\item{description}{The description string in a "PH_RANGE" entry.}
}
\value{
A \code{data.table} with columns: proteinID, description, fieldInfo,
commentary and refID. The description column is the extracted pH range,
and the fieldInfo should be all NAs.
}
\description{
Expand the string into a \code{data.table}.
}
