% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-nomenclature.R
\name{ParseProtein}
\alias{ParseProtein}
\title{Parse a "PROTEIN" entry into list of lists.}
\usage{
ParseProtein(description)
}
\arguments{
\item{description}{The description string in a "PROTEIN" entry.}
}
\value{
A list of lists with sublists having three elements:
id, organism and reference.
}
\description{
Split an entry string into a list of lists. Elements could
be accessed by \code{Find} or using the names.
}
\examples{
x <- paste0(
"PR\\t#1# Cavia porcellus  (#1# SULT1A2 <1,2,6,7>) <1,2,6,7>\\n",
"PR\\t#2# Mus musculus <11,18,19>\\n")
brendaDb:::ParseProtein(x)
}
