% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-helper.R
\name{ParseProteinNum}
\alias{ParseProteinNum}
\title{Parse protein information strings or reference strings.}
\usage{
ParseProteinNum(x, type)
}
\arguments{
\item{type}{Either "protein" or "reference".}

\item{ref}{A string in the format of "#1#" or "#1,2,3#" or "<1,3>".}
}
\value{
A string, or a vector of strings of protein numbers.
}
\description{
Given a string like "#1,45,72#", parse into a character vector
of c("1", "45", "72").
}
\examples{
parse.protein.num("#1,2,3#")
# [1] "1,2,3"
parse.protein.num("<123>")
# [1] "123"

}
