% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-nomenclature.R
\name{ParseSynonyms}
\alias{ParseSynonyms}
\title{Parse a "SYNONYMS" entry.}
\usage{
ParseSynonyms(description)
}
\arguments{
\item{description}{The description string in a "SYNONYMS" entry.}
}
\value{
A \code{data.table} with three columns: id, synonym and reference.
}
\description{
Expand the string into a \code{data.table}.
}
\examples{
x <- paste0(
"SY\\t aldehyde reductase\\nSY\\t dehydrogenase, alcohol\\n",
"SY\\t#8,10,95,97,112,113,135# ADH1 (#10# isozyme <202>)\\n",
"\\t<156,172,202,215,228,252,282>\\n")
brendaDb:::ParseSynonyms(x)

}
