% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{QueryBrenda}
\alias{QueryBrenda}
\title{Query for a specific enzyme.}
\usage{
QueryBrenda(brenda, EC)
}
\arguments{
\item{brenda}{A \code{tibble} containing information from BRENDA.}

\item{EC}{A string of the EC number.}
}
\value{
A \code{brenda.entry} object.
}
\description{
Use a EC number to retrieve information from the BRENDA
\code{tibble} read in by \code{ReadBrenda()}.
}
\examples{
df <- ReadBrenda(system.file("extdata", "brenda_download_test.txt",
                          package = "brendaDb"))
QueryBrenda(brenda = df, EC = "1.1.1.1")

}
\seealso{
ReadBrenda InitBrendaEntry
}
