% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{QueryBrendaBase}
\alias{QueryBrendaBase}
\title{Query for a specific enzyme.}
\usage{
QueryBrendaBase(brenda, EC, fields = F)
}
\arguments{
\item{brenda}{A \code{tibble} containing information from BRENDA.}

\item{EC}{A string of the EC number.}

\item{fields}{A character vector indicating fields to parse. Default is
FALSE, which would be returning all fields.}
}
\value{
A \code{brenda.entry} object.
}
\description{
Use a EC number to retrieve information from the BRENDA
\code{tibble} read in by \code{\link[=ReadBrenda]{ReadBrenda()}}.
}
\examples{
df <- ReadBrenda(system.file("extdata", "brenda_download_test.txt",
                          package = "brendaDb"))
brendaDb:::QueryBrendaBase(brenda = df, EC = "1.1.1.1")

}
\seealso{
\code{\link[=ReadBrenda]{ReadBrenda()}} \code{\link[=InitBrendaEntry]{InitBrendaEntry()}}
}
