% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.brenda}
\alias{read.brenda}
\title{Read BRENDA text file into matrix.}
\usage{
read.brenda(filepath)
}
\arguments{
\item{filepath}{A string indicating the path to the text file.}
}
\value{
A matrix containing information about the EC entries.
}
\description{
For each EC entry, split the annotations into three columns:
\itemize{
\item ID: EC number, e.g. 1.1.1.1
\item field: the content of the information, e.g. protein, localization
\item description: everything else
}
}
\examples{
df <- read.brenda("inst/extdata/brenda_download_test.txt")
# Reading BRENDA text file...
# Converting text into matrix...This might take a while...
head(df)
# ID         field             description
# 1.1.1.10   PROTEIN           PR	PR  #1# Cavia porcellus   (#1# SULT1A2...
# 1.1.1.10   RECOMMENDED_NAME  RN  L-xylulose reductase
# 1.1.1.10   SYSTEMATIC_NAME   SN  xylitol:NADP+ 4-oxidoreductase (L-xyl...

}
